backend/requirements.txt
flask
flask-cors
faiss-cpu
sentence-transformers
numpy
python-dotenv
PyPDF2
pandas
openpyxl